Feat : 채팅 말풍선 생성 함수 분리
